\name{RdpackTester-package}
\alias{RdpackTester-package}
\alias{RdpackTester}
\docType{package}
\title{
\packageTitle{RdpackTester}
}
\description{
\packageDescription{RdpackTester}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RdpackTester}
%\packageIndices{RdpackTester}

Some of the references here are non-existent or a modified to
test/demonstrate features of the citation macros.

\strong{Note:} Reference \insertCite{dummyArticle;textual}{Rdpack}
does not exist. It is a test that simple math in BibTeX entries works.


  \insertCite{nadler_diffusion_2005}{Rdpack}
  \insertCite{cayton_robust_2006}{Rdpack}
  
  \insertCite{fukunaga_introduction_1990}{Rdpack}
  \insertCite{fukunaga_introduction_1990_fake_vol}{Rdpack}
  \insertCite{fukunaga_introduction_1990_fake_number}{Rdpack}
  \insertCite{fukunaga_introduction_1990_fake_vol_number}{Rdpack}

}
\author{
\packageAuthor{RdpackTester}

Maintainer: \packageMaintainer{RdpackTester}
}
\references{
  \insertAllCited{}

  ---
  
Formulas are handled in fields were they make sense, here title,
journal, note:  

\insertRef{dummyArticle}{Rdpack}
  

}
\keyword{ package }
%\seealso{
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
\examples{
test_1()
}
